\name{andy}
\alias{andy}
\docType{data}
\title{
A sample of hamburger franchises in 75 cities from Big Andy's Burger Barn.
}
\usage{data(andy)}
\format{
  A data frame with 75 observations on the following 3 variables.
  \describe{
    \item{\code{sales}}{Monthly sales revenue ($1000s)}
    \item{\code{price}}{A price index for all products sold in a given month.}
    \item{\code{advert}}{Expenditure on advertising ($1000s)}
  }
}

\source{
http://www.principlesofeconometrics.com/poe4/data/dat/andy.dat

http://www.principlesofeconometrics.com/poe4/data/def/andy.def
}
\references{
Principles of Econometrics, Fourth Edition

by R. Carter Hill, William E. Griffiths and Guay C. Lim
}