\name{bangla}
\alias{bangla}
\docType{data}
\title{
bangla
}
\usage{data(bangla)}
\format{
  A data frame with 34 observations on the following 2 variables.
  \describe{
    \item{\code{p}}{Price of sugar cane divided by the price of jute
               (jute is an alternative crop to sugar cane, planted by Bangladesh farmers)}
    \item{\code{a}}{area of sugar cane planted in thousands of hectares in a region of Bangladesh}
  }
}

\source{
http://www.principlesofeconometrics.com/poe4/data/dat/bangla.dat

http://www.principlesofeconometrics.com/poe4/data/def/bangla.def
}
\references{
Principles of Econometrics, Fourth Edition

by R. Carter Hill, William E. Griffiths and Guay C. Lim
}