\name{food}
\alias{food}
\docType{data}
\title{
Food Expenditure and Income
}
\usage{data(food)}
\format{
  A data frame with 40 observations on the following 2 variables.
  \describe{
    \item{\code{food_exp}}{weekly food expenditure in $}
    \item{\code{income}}{weekly income in $100}
  }
}

\source{
http://www.principlesofeconometrics.com/poe4/data/dat/food.dat

http://www.principlesofeconometrics.com/poe4/data/def/food.def
}
\references{
Principles of Econometrics, Fourth Edition

by R. Carter Hill, William E. Griffiths and Guay C. Lim
}