\name{POE4}
\alias{POE4}
\docType{package}
\title{
A supplementary package for the 4th Edition of Principles of Econometrics by Hill, Griffiths, and Lim.
}
\description{
This packages serves as a convenience for accessing the datasets used in the textbook. This packages also contains the supplementary guide for reproducing the results found in the textbook.
}
\details{
\tabular{ll}{
Package: \tab POE4\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-09-26\cr
License: \tab MIT\cr
}
}
\author{
Nicole White

Maintainer: <nicole.m.white@outlook.com>
}
\references{
Principles of Econometrics, Fourth Edition by R. Carter Hill, William E. Griffiths and Guay C. Lim

http://principlesofeconometrics.com/poe4/poe4.htm
}