\name{table2_2}
\alias{table2_2}
\docType{data}
\title{
Food Expenditure and Income - Repeated Sampling
}
\usage{data(table2_2)}
\format{
  A data frame with 40 observations on the following 11 variables.
  \describe{
    \item{\code{x}}{weekly income in $100 from \code{\link{food}}}
    \item{\code{y1-y10}}{weekly food expenditure in $ for 10 households}
  }
}

\source{
http://www.principlesofeconometrics.com/poe4/data/dat/table2_2.dat

http://www.principlesofeconometrics.com/poe4/data/def/table2_2.def
}
\references{
Principles of Econometrics, Fourth Edition

by R. Carter Hill, William E. Griffiths and Guay C. Lim
}